Fix CDC upload, debounce, GPIO conflicts, and GUI improvements
Firmware:
- Fix CDC upload: spin-wait for TX busy, defer flash save to main loop
- Gate debug output on streaming mode to prevent protocol contamination
- Runtime debounce: read m_config.debounce_ms instead of compile-time constant
- Guard P0.06 output writes with BOARD_XIAO_NRF52840 (conflicts with F2M on Twiddler4)
- Disable UART log backend (TX pin 6 = F2M button, was driving it as output)
- Set REGOUT0 to 3.3V (chip erase resets to 1.8V default)
- Add flash storage module for persistent config

GUI:
- Move chord upload to background thread to prevent UI freeze
- Debounce slider config changes (0.2s) to avoid serial flooding
- Defer config sends to Save to Flash (prevents stream corruption)
- Stop/restart stream around Save to Flash and Reset Defaults
- Route chord events to Learn tab only when active (no stray hints)
- Fix double tab highlight on startup
- Add exercise progression: accuracy-based char/combo/lesson advancement
- Add chars dropdown, auto-start on first chord input
- Fix chord detection and hints when exercise prepared but not running

+++ firmware/include/nchorder_buttons.h
@@ -1,8 +1,8 @@
 /**
  * Northern Chorder - Button Scanning Driver
  *
- * GPIO-based button input handling with interrupt-driven detection
- * All 16 buttons are on GPIO Port 0 with internal pull-ups (active-low)
+ * GPIO-based button input handling with polling detection
+ * Twiddler 4: 19 buttons (4 thumb + 5x3 finger) with internal pull-ups (active-low)
  */
 
 #ifndef NCHORDER_BUTTONS_H
@@ -16,16 +16,15 @@
  * Button callback type
  * Called when button state changes (after debouncing)
  *
- * @param button_state 16-bit bitmask of currently pressed buttons
- *                     Bit 0 = T1 (N), Bit 1 = F1L, etc.
+ * @param button_state 32-bit bitmask of currently pressed buttons
+ *                     Bit 0 = T1, Bit 1 = F1L, ... Bit 16 = F0L, etc.
  *                     1 = pressed, 0 = released
  */
-typedef void (*buttons_callback_t)(uint16_t button_state);
+typedef void (*buttons_callback_t)(uint32_t button_state);
 
 /**
  * Initialize button GPIO pins
- * Configures all 16 button pins as inputs with pull-ups
- * Sets up GPIOTE for interrupt-driven detection
+ * Configures all button pins as inputs with pull-ups
  *
  * @return 0 on success, error code on failure
  */
@@ -33,12 +32,12 @@ uint32_t buttons_init(void);
 
 /**
  * Scan current button state
- * Reads all 16 GPIO pins and returns active-high bitmask
+ * Reads all GPIO pins and returns active-high bitmask
  * (GPIO is active-low, this function inverts for convenience)
  *
- * @return 16-bit bitmask of pressed buttons (1 = pressed)
+ * @return 32-bit bitmask of pressed buttons (1 = pressed)
  */
-uint16_t buttons_scan(void);
+uint32_t buttons_scan(void);
 
 /**
  * Register callback for button state changes
@@ -62,6 +61,6 @@ bool buttons_any_pressed(void);
  * @param bitmask Button state bitmask
  * @return Human-readable string like "T1+F1M+F2R"
  */
-const char* buttons_to_string(uint16_t bitmask);
+const char* buttons_to_string(uint32_t bitmask);
 
 #endif // NCHORDER_BUTTONS_H
